Call non-template base methods through 'this' pointer
From https://gcc.gnu.org/onlinedocs/gcc-9.3.0/gcc/Name-lookup.html:
"The C++ standard prescribes that all names that are not dependent on
template parameters are bound to their present definitions when parsing
a template function or class.5 Only names that are dependent are looked
up at the point of instantiation.
...
To make the code valid either use this->f(), or Base<T>::f()"

It is best to use the this->f() solution because it could be a virtual
method.

Also use the type-safe load() and store() methods where possible.

Bug: b/155302798
Change-Id: Id636bfb0b1aae9529bf7f3f58213ec7823d9eddd