Unconditionally define the RValue copy constructor
Previously we relied on the default copy constructor to be generated
when ENABLE_RR_DEBUG_INFO is disabled. This is not necessary as we can
just use the straightforward manual definition of it in either case.
Note RR_DEBUG_INFO_EMIT_VAR() is a no-op when ENABLE_RR_DEBUG_INFO is
disabled.

Bug: b/155302798
Change-Id: I04c020a795b4333e3c3d0a7ed375d21d9382e3d1