Don't inline rr::Variable methods
This is not a template class, so its methods don't have to be inlined.
Nor are they trivial enough to save us anything in performance or
binary size if they were inlined. Instead, this could reduce binary
size a bit because there are a large number of rr::Variable uses.

Bug: chromium:1086501
Change-Id: I1a6d078628ed7e634e0d90d6d8455e7d2938d134