Add decoder for BC6h
BC6h is the last BCn texture format we needed to enable the
textureCompressionBC flag.

The main design choice in this CL is how to decode all 14 modes in BC6h.
I initially tried to use a table that described the location of each
bitfield within the block. Transcribing the data from the block description
table proved to be error prone and I believed it would be difficult to debug.

My next attempt was to just use a switch statement that covered all 14
modes. Each case statement would decode each bit field within
the block based on the block description. This was perhaps the easiest
method to write, but it was overly verbose, taking up 400 lines of the
file. I felt that it was too much to take in.

And while writing that switch statement, I changed how I was getting
each bitfield. Instead of accessing bits N through M in the block, I
consumed the lowest (M-N) bits, shifting the next lowest bit into the
LSB of the block. This simplified a lot of logic and made it so I didn't
have to track how far into the block I was.

Next I'll try to implement the table idea again, but using the same
block description pattern I used when decoding with ConsumeBits.

Bug: b/151203718
Change-Id: I5dacbc07bb54ff4fc384db974feaf7cfd0055e96