Implement alternative output types for rr::Print
With this change, you can now modify Reactor.cpp to set the new
RR_PRINT_OUTPUT_TYPE macro to one of:

RR_PRINT_OUTPUT_TYPE_STDOUT 1   /* Print to stdout */
RR_PRINT_OUTPUT_TYPE_DEBUGGER 2 /* Print to debugger (OutputDebugString on Windows) */
RR_PRINT_OUTPUT_TYPE_STUB 3     /* Do not print, but call to function is emitted */

The 'stdout' option is the default. The 'debugger' option is currently
useful on Windows when it is difficult to capture stdout (e.g. debugging
Chrome), so this routes the output to the Visual Studio Debugger output
window while attached.

The last one, 'stub', is useful in combination with
ENABLE_RR_EMIT_PRINT_LOCATION in that debugging the code-gen assembly
allows you to inspect the string argument to rr::DebugPrintf and see
exactly where you are in Reactor code without paying the performance
cost of actually printing.

Bug: b/158678105
Change-Id: Ie1bd85888405e189200c0e5f8f3449551327e0ab