Implement alternative output types for rr::Print
With this change, you can now modify Reactor.cpp to set the new
RR_PRINT_OUTPUT_TYPE macro to one of:

RR_PRINT_OUTPUT_TYPE_STDOUT 1 /* Print to stdout, and to debugger if available (OutputDebugString on Windows). */
RR_PRINT_OUTPUT_TYPE_STUB 2   /* Do not print, but call to function is emitted */

The 'stdout' option is the default, and now also outputs to the debugger
on Windows (via OutputDebugString), which is useful when it is difficult
to capture stdout (e.g. debugging Chrome).

The 'stub' option is useful in combination with
ENABLE_RR_EMIT_PRINT_LOCATION in that debugging the code-gen assembly
allows you to inspect the string argument to rr::DebugPrintf and see
exactly where you are in Reactor code without paying the performance
cost of actually printing.

Bug: b/158678105
Change-Id: Ie1bd85888405e189200c0e5f8f3449551327e0ab