VulkanBenchmarks: add texture sampling to Triangle benchmarks
This change was originally started by capn@, which added texture
sampling to the Triangle benchmarks. I finished up this change with the
following changes and fixes:

* Implement a correct fill texture to white using copyBufferToImage. *
Add Buffer class to wrap vk::Buffer and vk::DeviceMemory. * Add support
for using a VK_EXT_headless_surface Window, which is the default on
non-Windows platforms for now. On Windows, can be forced to use this by
defining USE_HEADLESS_SURFACE = 1. * Fix out of bounds access because
renderPassBeginInfo.clearValueCount was set to 2 instead of 1. * Fix
illegal access to destructed ArrayRef in call to
commandBuffers[i].bindVertexBuffers. * Make it build and run on Linux
(fix compilation errors and use aforementioned headless surface mode)

Bug: b/176981107
Change-Id: Ifb8ac6f4488ef16fe21cc5941e94bbfeac86f846