Update patch set 6
Patch Set 6: Verified+1

> Patch Set 4: Code-Review+2
> 
> IIRC, these blitter functions were there to support quadLayout

Exactly. But it was kind of ironic that quad-layout, which is supposed to improve performance, was going through this slow path in the Blitter.

If and when we bring it back, we'll have benchmarks and a good baseline to compare it against.

Patch-set: 6
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Verified=+1