Optimize texel read operations
- When shifting the lowest element to the highest bits, we don't have to
mask the bits shifted in since they're already 0.
- When shifting the highest elements to the lowest bits, we don't have
to maks the bits shifted in when using an unsigned shift.
- For UNORM formats we don't have to perform integer shifts, as we can
adjust the magnitude as part of the normalization factor. Note we don't
do this for elements in the highest bits, because unsigned integer to
floating-point conversion does not have efficient instructions on x86
until AVX-512.

Bug: b/146579770
Change-Id: I714478b7ffde7a2b496f726afe7e64f2ad471c4d