Fix image slice / plane name usage
Image planes have a distinct meaning in Vulkan, signifying the disjoint
memory regions which store components of a texel format. Typically this
includes YCbCr formats that can have 2 or 3 planes for luminance and
chroma data.

Slices refer to 2D sections of 3D images. Currently we also use slices
for storing each sample of a multisample image separately. Note only
2D images can be multisampled, so there are no 3D images with slices
for the depth and the number of samples.

Bug: b/159210008
Change-Id: I4b70a7b8bb49e79c7494e28aa8280752a5b9727d