Update patch set 34
Patch Set 34:

(9 comments)

It took me a while to get through this change. It's a big one, for sure. It's hard for me to evaluate the merits of this change at a higher level as I admittedly am not familiar with this part of the code, but I did my best to provide useful feedback. From what I can tell, this change certainly does add complexity, as caching typically does, but I assume the performance benefits outweigh this increase in code complexity.

Patch-set: 34