Update patch set 39
Patch Set 39:

> Patch Set 34:
> 
> (9 comments)
> 
> It took me a while to get through this change. It's a big one, for sure. It's hard for me to evaluate the merits of this change at a higher level as I admittedly am not familiar with this part of the code, but I did my best to provide useful feedback. From what I can tell, this change certainly does add complexity, as caching typically does, but I assume the performance benefits outweigh this increase in code complexity.

It's not just performance, it's also correctness. Without this change, images (compressed or cubemaps) which get written to within a draw call or by a compute shader don't properly perform cube border updates and/or decompression. See these lines for examples of missing updates which would cause correctness issues (so not just performance issues).
https://swiftshader-review.googlesource.com/c/SwiftShader/+/45188/39/src/Device/Renderer.cpp#429
https://swiftshader-review.googlesource.com/c/SwiftShader/+/45188/39/src/Pipeline/ComputeProgram.cpp#302

There was currently no way of figuring out when an image would have been modified, since we only get a pointer to it's memory during descriptor set update or copy, which would then be used repeatedly, without the Image object having a way to know that its underlying memory was modified.

Patch-set: 39