Implement scatter/gather for 1- and 2-byte texels
Fixes out-of-bounds reads/writes and also avoids unaligned accesses.

This is a generic reference implementation. Future optimizations could
use unaligned 4-byte accesses if they're known to be safe (e.g. due to
padding) and efficient. We can also eliminate if we know we're in a
basic block post-dominated by the entry block and the number of
active invocations is a multiple of the SIMD width.

Bug: b/160531165
Change-Id: I892cfd3c7da8d8891cabe80695e5f35c57da73b4