Remove dead ASTC code for GLES
Surface::decodeASTC() was empty, and ENABLE_ASTC hard-coded to 0, so we
never actually supported ASTC for OpenGL ES.

We won't need it in the future since this legacy code is deprecated and
set to be replaced by SwANGLE (which supports ASTC LDR at the moment).
Also note ASTC LDR doesn't become mandatory until OpenGL ES 3.2, and
this legacy code lacks compute shaders to even support ES 3.1.

Bug: b/147536183
Change-Id: I61fe83214cfebe9f83bd49334474e095631af3e7