Adjust subTexelPrecisionBits
According to the Vulkan spec:

"subTexelPrecisionBits is the number of bits of precision in the
 division along an axis of an image used for minification and
 magnification filters. 2^subTexelPrecisionBits is the actual number
 of divisions along each axis of the image represented. Sub-texel
 values calculated during image sampling will snap to these locations
 when generating the filtered results."

The value currently defined for subTexelPrecisionBits (4) does not
represent accurately the precision that we use for filtering. We
either use 16 bit integers or 32 bit floating point computation
when we perform linear filtering, which both use more than 4 bits of
precision.

This was causing a few tests to fail as a few corner cases were
accumulating too much error for the tests to pass.

All these factors combined would cause failures:
- Non power of 2 textures, where test coordinates don't get
  subdivided to powers of 2 (so texture sizes like 31x55 or 57x35)
- Textures with large differences between neighboring pixels. Tests
  generally use gradients, so using the repeat mode was the only mode
  creating sharp differences between neighboring pixels, at the edges

In order to more accurately represent reality, subTexelPrecisionBits
was increased from 4 to 8.

Bug: b/151214291
Change-Id: I15810068a3e53f3812e07fbc5b04701159289350