Fix LLVM-10 CMake build
Some of the files for LLVM 10 were conditionally compiled for Debug
builds; however, two of these were also in the main source list:

${LLVM_DIR}/lib/Analysis/RegionInfo.cpp
${LLVM_DIR}/lib/Support/Debug.cpp

This lead to linker errors for missing symbols when, for e.g.,
Support/Debug.cpp was built because Support/circular_raw_ostream.cpp
was not compiled.

I decided to just fold in these few Debug-only sources into the main
list to keep it simple and similar to how LLVM 7 was being built.

Bug: b/152339534
Change-Id: Ie4dda85b3323f5036d29fd80908b1925b6c86853