Enable occlusionQueryPrecise
Swiftshader already counts all the fragments for this feature, it just
needed to be enabled in VkPhysicalDevice.cpp. The occlusion_query tests
that remain unsupported require VK_EXT_host_query_reset. This is
unsupported in SwiftShader as of now, though it is required for Vulkan
1.2 conformance.

Bug: b/148241855
Tests: dEQP-VK.query_pool.occlusion_query.*
Change-Id: Ib03df4225ed49865c648368ebf0547bfc3d117f7