Don't use 2D coordinates for 1D sampling
1D images were being treated as 2D ones with the second coordinate set
to 0.0. It required other workarounds like moving the array layer
coordinate and providing an addressing mode for the second coordinate.
These have been removed, and the texel address calculation now ignores
the second coordinate for 1D images.

Note this doesn't specialize filtering for the 1D case yet, so the
weights for filtering in the v-direction are explicitly zeroed out.

Fixes: b/162858741
Bug: b/134669567
Change-Id: I4684ee0337ac500860e9939087045a953bad813a