Specialize 1D sampling LOD calculation
Instead of reusing the 2D mipmap LOD calculation, this change adds a
separate method for computing the LOD for 1D sampling operations.

This also revealed that we would perform the anisotropic filtering
calculations on 1D sampling operations if enabled, which makes no sense.
This change limits it to 2D sampling only.

Bug: b/134669567
Change-Id: I05dd3fe358eccfcd01ebcff35872f56c6f0b0b70