Fix for dEQP-VK.wsi.*.swapchain.acquire.too_many_timeout
With the assertion for these tests fixed here:
https://gerrit.khronos.org/#/c/6075/

The *.too_many test passed locally, but not the *.too_many_timeout
test. The reason is that, when a timeout is provided, the acceptable
return value for failing to acquire an image is VK_TIMEOUT, not
VK_NOT_READY. Returning VK_TIMEOUT when a timeout is provided solves
the issue.

Bug: b/162249856
Change-Id: Id2733fec8422aad93a1aa351b481691c69692168
Tests: dEQP-VK.wsi.*.swapchain.acquire.*