Control Flow Integrity (cfi) fix
The cfi bot is throwing an error because an object is being assigned
with memory containing an invalid vtable. This is caused by assigning
a chunk of memory to an object before calling placement new using that
memory.

This cl assigns the pointer after the object has been constructed to
avoid triggering this error.

Bug: chromium:1116053
Change-Id: I541d7dc0e884aed7470b97d7d1d91a9bb67edb2d