Refactor image view extent retrieval
Image views are only accessed by shaders, and we've previously ensured
descriptors only have one field for holding either the depth of a 3D
texture or the array layer count. This value is accessed through the
getDepthOrLayerCount() method, and so the getMipLevelExtent() of the
ImageView class only has to return the 2D extent.

Bug: b/162315264
Change-Id: I93dd5bbabcd7efc8aa174beb38204e900c8be614