Add unit test for 11/10-bit minifloat denormal underflow
The reference code for conversion from 32-bit floating-point to 11- and
10-bit minifloat formats supports producing denormals for values smaller
than what can be represented as normalized representations. The
arithmetic can underflow to produce zero for values too small to be
represented as denormals.

This arithmetic contains a 32-bit shift operation which can shift by an
amount greater than 32, which has undefined behavior in C++ but produces
zero on x86 processors.

To fix the UB while retaining the currently intended behavior, this
change adds unit tests for the cutoff between the smallest denormal, and
zero.

Bug: b/147900455
Bug: chromium:1117433
Change-Id: Ic5e495dd822231d52a5551ee12733a616728d486