Make multisample resolve a Blitter method
The Vulkan spec explicitly states that "vkCmdBlitImage must not be used
for multisampled source or destination images. Use vkCmdResolveImage for
this purpose." And the only other way to obtain resolve multisample
results is by using resolve attachments as part of a subpass.

This split between blit operations and resolve operations should be
reflected by the Blitter interface so we have less confusion about
its blit() method being used to perform resolves. It will also
facilitate adding a fast path for common resolve operations.

Bug: b/147802090
Change-Id: I2549a5e7acd7ef9ec3f70f8ceb88ff5fc65a0d17