Drop the float literal suffix
In GLSL ES 3.00, float literals are allowed to have an 'f' or 'F'
suffix, but they don't change the parsing process since they're always
interpreted as single-precision IEEE-754 floating-point values.

The std::istringstream used for parsing the literal's string can have
implementation dependent behavior depending on the presence of the
suffix, so this change removes it. This matches what ANGLE does.

Fixes: b/168250854
Change-Id: I203833a2e817ca698894db5669941fb435b97868