SpirvShaderDebugger: Implement DebugValue
This debug instruction partially updates a local variable with SSA values.

This is typically used to update a `DebugLocalVariable` of a composite type, which holds structure member offsets from a base address.
To handle these, we allocate shadow memory to hold a copy of the entire variable in contiguous memory and have the `DebugLocalVariable` point to this memory. Whenever we encounter a `DebugValue`, we copy the necessary fields to the
shadow memory.

This approach is simplier and more performant than attempting to fiddle with `vk::dbg::Value`s for each `DebugValue`.

Bug: b/148401179