Vulkan/Debug: Various fixes / improvements to Thread
Remove the `Context::Lock` parameter from `enter()`. We don't want to force the lock to be held for the entire duration of the call.

Add `UpdateFrame` type alias.

Add the `UpdateFrame` parameter to `enter()` - you usually want to update the frame as soon as you've pushed the stack.

Switch `pauseAtFrame` from a `shared_ptr` to a `weak_ptr`.
If the frame is released before the thread finds it to pause, then the thread will instead pause at the next frame (following the logic of a null `pauseAtFrame`).

Fixed the assignment of `pauseAtFrame` in `stepOut()` - it was looking at the current frame, not the next one up.

Bug: b/145351270
Change-Id: I1cb85af58c666a7793145480abc95a4ec82e1859