Fix clamping cube face coordinates to [0.0, 1.0) range
While cube face coordinates are nominally already in the [0.0, 1.0]
range due to the projection of the sampling coordinates onto the cube
face, the projection doesn't result in such normalized cube face
coordinates for Inf and NaN values in the sampling coordinates.

This change causes us to always clamp, matching the legacy OpenGL ES
code now.

Bug: b/132981873
Change-Id: Ib89d032ed4c5945f115269a4e8989aa03e52a6b5