WIP-optimize-cube-coordinate-normalization
This change forces each coordinate to the [0.0, 1.0) range with a single
unsigned integer Max() operation, as opposed to a range clamp using
Min/Max operations. This results in mapping negative values to 1.0
(or rather 1.0 minus 1 ULP) instead of 0.0, but this is fine since
-Inf and -NaN are the only expected negative input values, and they
result from meaningless sampling coordinates.

Note that this assumption about negative values requires the projection
division to produce 1.0 for x / x, which isn't guaranteed by the
reciprocal() function unless explicitly requested.

Bug: b/169802633
Change-Id: Ic97bc59fa000b679b7e7e8491c1c31e55808aa9a