Revert sampling parameter initialization workaround
1D image sampling used to be treated as 2D image sampling with the
second coordinate being ignored. But a bug was causing us to use the
second component of the 'offset' parameter. This was worked around by
initializing all parameters provided to the sampling routine.

This costs some performance, and since we've fixed the 'offset' bug and
have specialized handling of 1D images now, the workaround can be
reverted.

Change-Id: Id69b68465d7755df35fe12619957971681e4c0b7
Fixes: b/136149446
Bug: b/134669567