Fix Subzero arithmetic ops with two constant args
Detect when both args are const, and convert the lhs arg to a variable
as Subzero expects. Note that this fixes a failure that was only caught
at optimization level Om1 (no optimizations) when targeting x86. With
optimizations enabled, Subzero is able to optimize the operations and
bypasses the failure path, which explains why this hasn't been an issue
before.

Wrote unit tests to validate one const arg on either side, and two const
args.

Bug: tbd
Change-Id: If777ca4d4df189b1b7fd67ffe572044c8d0508c0