Move interpolant clamping out of interpolate()
Vulkan never clamps interpolants, except for depth values (specifically
for fixed-point formats, and for floating-point formats when the
unrestricted depth extension is not enabled).

This change moves the clamping operation out of interpolate(), which
makes it easier to see depth clamping being performed when necessary,
and will facilitate changes in applying depth bias.

Bug: b/139341727
Change-Id: I8cef29d63fef00308ef2f813b4feb687c83baa86