Make abort on LOG_TRAP opt-in
A developer can change SWIFTSHADER_LOG_LEVEL to Debug or Verbose to make
LOG_TRAP abort when a debugger is attached. This keeps debugging easy in
most circumstances where devs don't care about unsupported struct
extensions. But it still allows devs to go right to where an unsupported
structure is being used if they so desire.

Bug: b/169868497
Change-Id: I916506ee557f8de01516fe4efb33044c2367db77