Subzero: fix Ice::BitVector::grow not copying old to new data
UBSAN pointed out that memcpy cannot have nullptr dest or source,
which I've also fixed; however, this also made me realize a latent bug:
the dest and source args needed to be swapped. It seems this bug hasn't
bit us yet because the normal use-case is to resize the bitvector, and
then call 'set' to set all bits to 0 or 1. In any case, still worth
fixing.

Bug: b/171581990