Implement VK_KHR_shader_float_controls
This extension allows users to query and specify how the implementation
handles floats. It includes several SPIR-V capabilities to support
this behavior.

Part of this extension requires making denorms flush to zero. So
CPUID::setFlushToZero has been expanded to work on the x86 architecture
in general, rather than only on windows. Many of our shader intrinsics
fail to properly flush denorms. So in those cases we manually flush any
intrinsic that fails a denorm flush dEQP test.

Because this change requires the latest SPIR-V version as of Vulkan 1.2,
VkConfig.hpp has been updated to include the current version of SPIR-V
that the rest of the code refers back to.

Bug: b/169904252 b/171498580
Tests: dEQP-VK.spirv_assembly.instruction.compute.float_controls.*
Change-Id: I2926f015bdccad0ca2a4cfae195eeeba09781f07