Add script to verify proper Memset template use
Since there does not seem to be any elegant way of
checking whether a class is the first base class in
c++ (given that the 2009 c++ proposal n2965 was
rejected), this python script does it by scanning
all source files for illegal patterns.

The python file was implemented in a way to enable
easy addition of further checks, should need arise.

Bug: b/152253831
Change-Id: I25701a3302579700ccb6222141843d27f38d9069