Fix LLVM alloca array size type
https://llvm.org/docs/LangRef.html#alloca-instruction does not specify
the type of the 'NumElements' parameter explicitly, but it is multiplied
by sizeof(<type>), which itself is of type size_t and thus 64-bit on
a 64-bit platform. This assumption is made by MemorySanitizer when it
emits instructions corresponding to this multiplication, which get
constant folded and requires both operands to be of the same bit size.

Bug: b/155148722
Change-Id: I432e69bfcea4d40ce7b593637fd242bc1a25a714